Changed all CreateTime properties to long?s

+++ Cisco.DnaCenter.Api/Data/FlowAnalysisListOutputResponse.cs
@@ -53,7 +53,7 @@ namespace Cisco.DnaCenter.Api.Data
 		/// Gets or Sets CreateTime
 		/// </summary>
 		[DataMember(Name = "createTime", EmitDefaultValue = false)]
-		public int? CreateTime { get; set; }
+		public long? CreateTime { get; set; }
 
 		/// <summary>
 		/// Gets or Sets DestIP
